Cleanup

+++ saturnin/src/video/gui.cpp
@@ -1875,10 +1875,8 @@ void showDebugTexturesWindow(core::EmulatorContext& state, bool* opened) {
                     const auto& key        = keys_list[current_texture_idx].second;
                     const auto  opengl_tex = state.opengl()->getOpenglTexture(key);
 
-                    // auto tex_id = state.opengl()->generateTextureFromTextureArrayLayer(opengl_tex->texture_array_index, key);
-                    auto tex_id = state.opengl()->generateTextureFromTextureArrayLayer(video::GuiTextureType::layer_buffer, key);
-                    // auto tex_id = state.opengl()->getRenderedBufferTextureId(video::GuiTextureType::layer_buffer);
-                    //  u32 tex_id = 0; // :FIXME:
+                    auto tex_id = state.opengl()->generateTextureFromTextureArrayLayer(video::GuiTextureType::layer_buffer,
+                                                                                       static_cast<u8>(key));
                     ImGui::SetCursorPos(layer_window_pos);
                     const auto child_size = ImVec2(area_3_width, ImGui::GetContentRegionAvail().y);
                     ImGui::BeginChild("ChildTextureLayer", child_size, false, window_flags);
